Refactor: split gre.rs into submodules under src/parsers/gre/
Split the 3,838-line monolithic gre.rs into focused submodules:
- mod.rs: try_parse dispatch, constants, GRE message extraction
- connect_resp.rs: ConnectResp payload builder
- game_state.rs: GameStateMessage/QueuedGameStateMessage payload builder
- game_result.rs: GameStage_GameOver detection and result payload
- annotations.rs: annotation extraction (ZoneTransfer, ObjectIdChanged)
- turn_info.rs: turn/phase info extraction from gameInfo.turnInfo
- helpers.rs: shared extraction utilities (extract_string_array, etc.)
- test_fixtures.rs: shared test body builders used across submodules

No public API changes. All 737 tests pass unchanged.

Closes manasight/manasight-docs#213
